corrige tipo de variável para soma e ajusta formatação da saída no exercício 2

+++ 1º Semestre/Programação I/Atividades While e For/src/exercicio_2.java
@@ -2,18 +2,18 @@ import java.util.Scanner;
 
 public class exercicio_2 {
     public static void main(String[] args) {
-        int soma = 0, num, qtd;
+        double soma = 0, num, qtd;
         double media;
         Scanner in = new Scanner(System.in);
 
         System.out.print("Escreva a quantidade de números que ele somara: ");
-        qtd = in.nextInt();
+        qtd = in.nextDouble();
 
         for (num = 1; num <= qtd; num++) {
             soma += num; // soma acumulativa
         }
-        media = (double) (soma) / num;
-        System.out.printf("\nA soma dos %d primeiros números é: %d\n", qtd, soma);
+        media = soma / num;
+        System.out.printf("\nA soma dos %.2f primeiros números é: %.2f\n", qtd, soma);
         System.out.printf("E a média é: %.2f", media);
     }
